переносы функций для экрана статистики в модуль статистики. удаление ненужных файлов, небольшая чистка кода
