+++ js/three-cards-game-screen-view.js
@@ -1,7 +1,6 @@
 import getFooterMarkup from './footer.js';
 import AbstractView from './abstract-view.js';
 import {getLevelProgressBar} from './utils.js';
-import enterNextLevel from './enter-next-level.js';
 import getHeader from './header.js';
 import {userState} from './utils.js';
