+++ js/utils.js
@@ -10,83 +10,6 @@ const createUserdata = () => {
   userState[`answers`] = [];
 };
 
-const getLevelProgressBar = (userAnswers) => {
-  let userLevels;
-  let answerType;
-  userLevels = userAnswers.map((answer) => {
-    if (!answer.solved || answer.time > AnswerTime.TIMEOUT) {
-      answerType = `wrong`;
-    } else if (answer.solved && answer.time > AnswerTime.SLOW && answer.time < AnswerTime.TIMEOUT) {
-      answerType = `slow`;
-    } else if (answer.solved && answer.time < AnswerTime.FAST) {
-      answerType = `fast`;
-    } else if (answer.solved && answer.time > AnswerTime.FAST && answer.time < AnswerTime.SLOW) {
-      answerType = `correct`;
-    }
-    let type = answerType;
-    return `<li class="stats__result stats__result--${type}"></li>`;
-  });
-  let userRemainedLevels = [];
-  for (let i = 0; i < 10 - userAnswers.length; i++) {
-    userRemainedLevels.push(`<li class="stats__result stats__result--unknown"></li>`);
-  }
-  userLevels.push(userRemainedLevels);
-  return userLevels;
-};
-
-const getFastAnswersValue = (arrayWithAnswers) => {
-  let fastAnswersValue = 0;
-  for (let i = 0; i < arrayWithAnswers.length; i++) {
-    if (arrayWithAnswers[i].time < AnswerTime.FAST && arrayWithAnswers[i].solved) {
-      fastAnswersValue++;
-    }
-  }
-  return fastAnswersValue;
-};
-
-const getSlowAnswersValue = (arrayWithAnswers) => {
-  let slowAnswersValue = 0;
-  for (let i = 0; i < arrayWithAnswers.length; i++) {
-    if (arrayWithAnswers[i].time > AnswerTime.SLOW && arrayWithAnswers[i].time < AnswerTime.TIMEOUT && arrayWithAnswers[i].solved) {
-      slowAnswersValue++;
-    }
-  }
-  return slowAnswersValue;
-};
-
-const getCompletedLevelsValue = (arrayWithAnswers) => {
-  let completedLevelValue = 0;
-  for (let i = 0; i < arrayWithAnswers.length; i++) {
-    if (arrayWithAnswers[i].solved) {
-      completedLevelValue++;
-    }
-  }
-  return completedLevelValue;
-};
-
-const getScore = (userResult, remainingLifes) => {
-  let userPoints = 0;
-  if (userResult.length < 10) {
-    return -1;
-  } else {
-    for (let i = 0; i < userResult.length; i++) {
-      let currentUser = userResult[i];
-      if (currentUser.solved) {
-        userPoints = userPoints + AnswerPoints.NORMAL;
-        if (currentUser.time < AnswerTime.FAST) {
-          userPoints = userPoints + AnswerPoints.BONUS;
-        } else if (currentUser.time > AnswerTime.SLOW && currentUser.time < AnswerTime.TIMEOUT) {
-          userPoints = userPoints - AnswerPoints.BONUS;
-        } else if (currentUser.time > AnswerTime.TIMEOUT) {
-          userPoints = userPoints - AnswerPoints.NORMAL;
-        }
-      }
-    }
-  }
-  userPoints = userPoints + (remainingLifes * AnswerPoints.BONUS);
-  return userPoints;
-};
-
 function createTimer(duration) {
   let timeRemain = duration;
   return {
@@ -104,22 +27,6 @@ function createTimer(duration) {
     time: timeRemain
   };
 }
-
-const getStatsTitle = (userAnswers, state) => {
-  if (getScore(userAnswers, state.lives) === -1) {
-    return `Проигрыш!`;
-  } else {
-    return `Победа!`;
-  }
-};
-
-const getStatsResult = (userAnswers, state) => {
-  if (getScore(userAnswers, state.lives) === -1) {
-    return `Fail!`;
-  } else {
-    return getScore(userAnswers, state.lives);
-  }
-};
 
 const getElementFromTemplate = (template) => {
   const pageElement = document.createElement(`div`);
@@ -203,4 +110,81 @@ const handleResultOfThreeCardsLevel = (evt, firstCardAnswer, secondCardAnswer, t
   return answerOnCard;
 };
 
-export {getElementFromTemplate, getScore, createTimer, getFastAnswersValue, getSlowAnswersValue, getCompletedLevelsValue, getLevelProgressBar, getStatsTitle, getStatsResult, createUserdata, userState, showGameScreen, handleResultOfTwoCardsLevel, handleResultOfOneCardLevel, handleResultOfThreeCardsLevel, findDifferentCard};
+const getScore = (userResult, remainingLifes) => {
+  let userPoints = 0;
+  if (userResult.length < 10) {
+    return -1;
+  } else {
+    for (let i = 0; i < userResult.length; i++) {
+      let currentUser = userResult[i];
+      if (currentUser.solved) {
+        userPoints = userPoints + AnswerPoints.NORMAL;
+        if (currentUser.time < AnswerTime.FAST) {
+          userPoints = userPoints + AnswerPoints.BONUS;
+        } else if (currentUser.time > AnswerTime.SLOW && currentUser.time < AnswerTime.TIMEOUT) {
+          userPoints = userPoints - AnswerPoints.BONUS;
+        } else if (currentUser.time > AnswerTime.TIMEOUT) {
+          userPoints = userPoints - AnswerPoints.NORMAL;
+        }
+      }
+    }
+  }
+  userPoints = userPoints + (remainingLifes * AnswerPoints.BONUS);
+  return userPoints;
+};
+
+const getFastAnswersValue = (arrayWithAnswers) => {
+  let fastAnswersValue = 0;
+  for (let i = 0; i < arrayWithAnswers.length; i++) {
+    if (arrayWithAnswers[i].time < AnswerTime.FAST && arrayWithAnswers[i].solved) {
+      fastAnswersValue++;
+    }
+  }
+  return fastAnswersValue;
+};
+
+const getSlowAnswersValue = (arrayWithAnswers) => {
+  let slowAnswersValue = 0;
+  for (let i = 0; i < arrayWithAnswers.length; i++) {
+    if (arrayWithAnswers[i].time > AnswerTime.SLOW && arrayWithAnswers[i].time < AnswerTime.TIMEOUT && arrayWithAnswers[i].solved) {
+      slowAnswersValue++;
+    }
+  }
+  return slowAnswersValue;
+};
+
+const getCompletedLevelsValue = (arrayWithAnswers) => {
+  let completedLevelValue = 0;
+  for (let i = 0; i < arrayWithAnswers.length; i++) {
+    if (arrayWithAnswers[i].solved) {
+      completedLevelValue++;
+    }
+  }
+  return completedLevelValue;
+};
+
+const getLevelProgressBar = (userAnswers) => {
+  let userLevels;
+  let answerType;
+  userLevels = userAnswers.map((answer) => {
+    if (!answer.solved || answer.time > AnswerTime.TIMEOUT) {
+      answerType = `wrong`;
+    } else if (answer.solved && answer.time > AnswerTime.SLOW && answer.time < AnswerTime.TIMEOUT) {
+      answerType = `slow`;
+    } else if (answer.solved && answer.time < AnswerTime.FAST) {
+      answerType = `fast`;
+    } else if (answer.solved && answer.time > AnswerTime.FAST && answer.time < AnswerTime.SLOW) {
+      answerType = `correct`;
+    }
+    let type = answerType;
+    return `<li class="stats__result stats__result--${type}"></li>`;
+  });
+  let userRemainedLevels = [];
+  for (let i = 0; i < 10 - userAnswers.length; i++) {
+    userRemainedLevels.push(`<li class="stats__result stats__result--unknown"></li>`);
+  }
+  userLevels.push(userRemainedLevels);
+  return userLevels;
+};
+
+export {getElementFromTemplate, createTimer, createUserdata, userState, showGameScreen, handleResultOfTwoCardsLevel, handleResultOfOneCardLevel, handleResultOfThreeCardsLevel, findDifferentCard, getLevelProgressBar, getScore, getFastAnswersValue, getSlowAnswersValue, getCompletedLevelsValue};
